Moves GameResult to new records module.
