+++ src/chesster/records/__init__.py
@@ -1,0 +1,3 @@
+"""The records module for Chesster"""
+from .result import GameResult
+
